split sct.utils in multiple modules
- sct.utils.fs: filesystem related stuff
- sct.utils.shell: user/shell convenience stull
- sct.utils.sys: other system/misc stuff
