--- spinalcordtoolbox/utils.py
@@ -1,712 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8
-# Collection of useful functions
-
-import io
-import sys
-import os
-import re
-import shutil
-import tempfile
-import datetime
-import logging
-import argparse
-import subprocess
-import time
-import shlex
-import atexit
-from enum import Enum
-
-import tqdm
-
-logger = logging.getLogger(__name__)
-
-__sct_dir__ = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
-__version__ = _version_string()
-__data_dir__ = os.path.join(__sct_dir__, 'data')
-__deepseg_dir__ = os.path.join(__data_dir__, 'deepseg_models')
-
-if os.getenv('SENTRY_DSN', None):
-    # do no import if Sentry is not set (i.e., if variable SENTRY_DSN is not defined)
-    import raven
-
-# TODO: add test
-def init_sct(log_level=1, update=False):
-    """
-    Initialize the sct for typical terminal usage
-    :param log_level: int: 0: warning, 1: info, 2: debug.
-    :param update: Bool: If True, only update logging log level. Otherwise, set logging + Sentry.
-    :return:
-    """
-    dict_log_levels = {0: 'WARNING', 1: 'INFO', 2: 'DEBUG'}
-
-    def _format_wrap(old_format):
-        def _format(record):
-            res = old_format(record)
-            if record.levelno >= logging.ERROR:
-                res = "\x1B[31;1m{}\x1B[0m".format(res)
-            elif record.levelno >= logging.WARNING:
-                res = "\x1B[33m{}\x1B[0m".format(res)
-            else:
-                pass
-            return res
-        return _format
-
-    # Set logging level for logger and increase level for global config (to avoid logging when calling child functions)
-    logger.setLevel(getattr(logging, dict_log_levels[log_level]))
-    logging.root.setLevel(getattr(logging, dict_log_levels[log_level]))
-
-    if not update:
-        # Initialize logging
-        hdlr = logging.StreamHandler(sys.stdout)
-        fmt = logging.Formatter()
-        fmt.format = _format_wrap(fmt.format)
-        hdlr.setFormatter(fmt)
-        logging.root.addHandler(hdlr)
-
-        # Sentry config
-        init_error_client()
-        if os.environ.get("SCT_TIMER", None) is not None:
-            add_elapsed_time_counter()
-
-        # Display SCT version
-        logger.info('\n--\nSpinal Cord Toolbox ({})\n'.format(__version__))
-
-
-def add_elapsed_time_counter():
-
-    class Timer():
-        def __init__(self):
-            self._t0 = time.time()
-
-        def atexit(self):
-            print("Elapsed time: %.3f seconds" % (time.time() - self._t0))
-    t = Timer()
-    atexit.register(t.atexit)
-
-
-def traceback_to_server(client):
-    """
-        Send all traceback children of Exception to sentry
-    """
-
-    def excepthook(exctype, value, traceback):
-        if issubclass(exctype, Exception):
-            client.captureException(exc_info=(exctype, value, traceback))
-        sys.__excepthook__(exctype, value, traceback)
-
-    sys.excepthook = excepthook
-
-
-def init_error_client():
-    if os.getenv('SENTRY_DSN'):
-        logger.debug('Configuring sentry report')
-        try:
-            client = raven.Client(
-                release=__version__,
-                processors=(
-                    'raven.processors.RemoveStackLocalsProcessor',
-                    'raven.processors.SanitizePasswordsProcessor'),
-            )
-            server_log_handler(client)
-            traceback_to_server(client)
-            old_exitfunc = sys.exitfunc
-
-            def exitfunc():
-                sent_something = False
-                try:
-                    # implementation-specific
-                    for handler, args, kw in atexit._exithandlers:
-                        if handler.__module__.startswith("raven."):
-                            sent_something = True
-                except:
-                    pass
-                old_exitfunc()
-                if sent_something:
-                    print("Note: you can opt out of Sentry reporting by editing the file ${SCT_DIR}/bin/sct_launcher and delete the line starting with \"export SENTRY_DSN\"")
-            sys.exitfunc = exitfunc
-        except raven.exceptions.InvalidDsn:
-            # This could happen if sct staff change the dsn
-            logger.debug('Sentry DSN not valid anymore, not reporting errors')
-
-
-def server_log_handler(client):
-    """ Adds sentry log handler to the logger
-
-    :return: the sentry handler
-    """
-    from raven.handlers.logging import SentryHandler
-
-    sh = SentryHandler(client=client, level=logging.ERROR)
-
-    # Don't send Sentry events for command-line usage errors
-    old_emit = sh.emit
-
-    def emit(self, record):
-        if record.message.startswith("Command-line usage error:"):
-            return
-        return old_emit(record)
-
-    sh.emit = lambda x: emit(sh, x)
-
-    fmt = ("[%(asctime)s][%(levelname)s] %(filename)s: %(lineno)d | "
-           "%(message)s")
-    formatter = logging.Formatter(fmt=fmt, datefmt="%H:%M:%S")
-    formatter.converter = time.gmtime
-    sh.setFormatter(formatter)
-
-    logger.addHandler(sh)
-    return sh
-
-
-class ActionCreateFolder(argparse.Action):
-    """
-    Custom action: creates a new folder if it does not exist. If the folder
-    already exists, do nothing.
-
-    The action will strip off trailing slashes from the folder's name.
-    Source: https://argparse-actions.readthedocs.io/en/latest/
-    """
-    @staticmethod
-    def create_folder(folder_name):
-        """
-        Create a new directory if not exist. The action might throw
-        OSError, along with other kinds of exception
-        """
-        if not os.path.isdir(folder_name):
-            os.mkdir(folder_name)
-
-        folder_name = os.path.normpath(folder_name)
-        return folder_name
-
-    def __call__(self, parser, namespace, values, option_string=None):
-        if type(values) == list:
-            folders = list(map(self.create_folder, values))
-        else:
-            folders = self.create_folder(values)
-
-        # Add the attribute
-        setattr(namespace, self.dest, folders)
-
-
-# TODO: Use for argparse wherever type_value=[['delim'], 'type'] was used
-def list_type(delimiter, subtype):
-    """
-        Factory function that returns a list parsing function, which can be
-        used with argparse's `type` option. This allows for more complex type
-        parsing, and preserves the behavior of the old msct_parser.
-    """
-    def list_typecast_func(string):
-        return [subtype(v) for v in string.split(delimiter)]
-
-    return list_typecast_func
-
-
-class Metavar(Enum):
-    """
-    This class is used to display intuitive input types via the metavar field of argparse
-    """
-    file = "<file>"
-    str = "<str>"
-    folder = "<folder>"
-    int = "<int>"
-    list = "<list>"
-    float = "<float>"
-
-    def __str__(self):
-        return self.value
-
-
-class SmartFormatter(argparse.ArgumentDefaultsHelpFormatter):
-    """
-    Custom formatter that inherits from HelpFormatter, which adjusts the default width to the current Terminal size,
-    and that gives the possibility to bypass argparse's default formatting by adding "R|" at the beginning of the text.
-    Inspired from: https://pythonhosted.org/skaff/_modules/skaff/cli.html
-    """
-
-    def __init__(self, *args, **kw):
-        self._add_defaults = None
-        super(SmartFormatter, self).__init__(*args, **kw)
-        # Update _width to match Terminal width
-        try:
-            self._width = shutil.get_terminal_size()[0]
-        except (KeyError, ValueError):
-            logger.warning('Not able to fetch Terminal width. Using default: %s'.format(self._width))
-
-    # this is the RawTextHelpFormatter._fill_text
-    def _fill_text(self, text, width, indent):
-        # print("splot",text)
-        if text.startswith('R|'):
-            paragraphs = text[2:].splitlines()
-            rebroken = [argparse._textwrap.wrap(tpar, width) for tpar in paragraphs]
-            rebrokenstr = []
-            for tlinearr in rebroken:
-                if (len(tlinearr) == 0):
-                    rebrokenstr.append("")
-                else:
-                    for tlinepiece in tlinearr:
-                        rebrokenstr.append(tlinepiece)
-            return '\n'.join(rebrokenstr)  # (argparse._textwrap.wrap(text[2:], width))
-        return argparse.RawDescriptionHelpFormatter._fill_text(self, text, width, indent)
-
-    # this is the RawTextHelpFormatter._split_lines
-    def _split_lines(self, text, width):
-        if text.startswith('R|'):
-            lines = text[2:].splitlines()
-            while lines[0] == '':  # Discard empty start lines
-                lines = lines[1:]
-            offsets = [re.match("^[ \t]*", l).group(0) for l in lines]
-            wrapped = []
-            for i in range(len(lines)):
-                li = lines[i]
-                if len(li) > 0:
-                    o = offsets[i]
-                    ol = len(o)
-                    init_wrap = argparse._textwrap.fill(li, width).splitlines()
-                    first = init_wrap[0]
-                    rest = "\n".join(init_wrap[1:])
-                    rest_wrap = argparse._textwrap.fill(rest, width - ol).splitlines()
-                    offset_lines = [o + wl for wl in rest_wrap]
-                    wrapped = wrapped + [first] + offset_lines
-                else:
-                    wrapped = wrapped + [li]
-            return wrapped
-        return argparse.HelpFormatter._split_lines(self, text, width)
-
-    def _get_help_string(self, action):
-        if action.default not in [None, "", [], (), {}]:
-            return super()._get_help_string(action)
-        else:
-            return action.help
-
-
-# Modified from http://shallowsky.com/blog/programming/python-tee.html
-class Tee:
-    def __init__(self, _fd1, _fd2):
-        self.fd1 = _fd1
-        self.fd2 = _fd2
-
-    # This is breaking pytest for test_sct_run_batch.py somehow.
-    # I think it is ok to omit this, allowing the fd objects to close themselves
-    # this prevents closing an fd in use elsewhere.
-    # def __del__(self):
-    #     self.close()
-
-    def close(self):
-        if self.fd1 != sys.__stdout__ and self.fd1 != sys.__stderr__:
-            self.fd1.close()
-        if self.fd2 != sys.__stdout__ and self.fd2 != sys.__stderr__:
-            self.fd2.close()
-
-    def write(self, text):
-        self.fd1.write(text)
-        self.fd2.write(text)
-
-    def flush(self):
-        self.fd1.flush()
-        self.fd2.flush()
-
-
-def abspath(fname):
-    """
-    Get absolute path of input file name or path. Deals with tilde.
-
-    '~/code/bla' ------------------> '/usr/bob/code/bla'
-    '~/code/bla/pouf.txt' ---------> '/usr/bob/code/bla/pouf.txt'
-    '/usr/bob/code/bla' -----------> '/usr/bob/code/bla'
-    '/usr/bob/code/bla/pouf.txt' --> '/usr/bob/code/bla/pouf.txt'
-
-    :param fname:
-    :return:
-    """
-    return os.path.abspath(os.path.expanduser(fname))
-
-
-def check_exe(name):
-    """
-    Ensure that a program exists
-
-    :param name: str: name or path to program
-    :return: path of the program or None
-    """
-    def is_exe(fpath):
-        return os.path.isfile(fpath) and os.access(fpath, os.X_OK)
-
-    fpath, _ = os.path.split(name)
-    if fpath and is_exe(name):
-        return fpath
-    else:
-        for path in os.environ["PATH"].split(os.pathsep):
-            path = path.strip('"')
-            exe_file = os.path.join(path, name)
-            if is_exe(exe_file):
-                return exe_file
-
-    return None
-
-
-def parse_num_list(str_num):
-    """
-    Parse numbers in string based on delimiter ',' or ':'
-
-    .. note::
-        Examples:
-        '' -> []
-        '1,2,3' -> [1, 2, 3]
-        '1:3,4' -> [1, 2, 3, 4]
-        '1,1:4' -> [1, 2, 3, 4]
-
-    :param str_num: str
-    :return: list of ints
-    """
-    list_num = list()
-
-    if not str_num:
-        return list_num
-
-    elements = str_num.split(",")
-    for element in elements:
-        m = re.match(r"^\d+$", element)
-        if m is not None:
-            val = int(element)
-            if val not in list_num:
-                list_num.append(val)
-            continue
-        m = re.match(r"^(?P<first>\d+):(?P<last>\d+)$", element)
-        if m is not None:
-            a = int(m.group("first"))
-            b = int(m.group("last"))
-            list_num += [x for x in range(a, b + 1) if x not in list_num]
-            continue
-        raise ValueError("unexpected group element {} group spec {}".format(element, str_num))
-
-    return list_num
-
-
-def parse_num_list_inv(list_int):
-    """
-    Take a list of numbers and output a string that reduce this list based on delimiter ';' or ':'
-
-    .. note::
-        Note: we use ; instead of , for compatibility with csv format.
-        Examples:
-        [] -> ''
-        [1, 2, 3] --> '1:3'
-        [1, 2, 3, 5] -> '1:3;5'
-
-    :param list_int: list: list of ints
-    :return: str_num: string
-    """
-    # deal with empty list
-    if not list_int or list_int is None:
-        return ''
-    # Sort list in increasing number
-    list_int = sorted(list_int)
-    # initialize string
-    str_num = str(list_int[0])
-    colon_is_present = False
-    # Loop across list elements and build string iteratively
-    for i in range(1, len(list_int)):
-        # if previous element is the previous integer: I(i-1) = I(i)-1
-        if list_int[i] == list_int[i - 1] + 1:
-            # if ":" already there, update the last chars (based on the number of digits)
-            if colon_is_present:
-                str_num = str_num[:-len(str(list_int[i - 1]))] + str(list_int[i])
-            # if not, add it along with the new int value
-            else:
-                str_num += ':' + str(list_int[i])
-                colon_is_present = True
-        # I(i-1) != I(i)-1
-        else:
-            str_num += ';' + str(list_int[i])
-            colon_is_present = False
-
-    return str_num
-
-
-def tmp_create(basename=None):
-    """Create temporary folder and return its path
-    """
-    prefix = "sct-%s-" % datetime.datetime.now().strftime("%Y%m%d%H%M%S.%f")
-    if basename:
-        prefix += "%s-" % basename
-    tmpdir = tempfile.mkdtemp(prefix=prefix)
-    logger.info("Creating temporary folder (%s)" % tmpdir)
-    return tmpdir
-
-
-def send_email(addr_to, addr_from, subject, message='', passwd=None, filename=None, html=False, smtp_host=None, smtp_port=None, login=None):
-    if smtp_host is None:
-        smtp_host = os.environ.get("SCT_SMTP_SERVER", "smtp.gmail.com")
-    if smtp_port is None:
-        smtp_port = int(os.environ.get("SCT_SMTP_PORT", 587))
-    if login is None:
-        login = addr_from
-
-    import smtplib
-    from email.mime.multipart import MIMEMultipart
-    from email.mime.text import MIMEText
-    from email.mime.base import MIMEBase
-    from email import encoders
-
-    if html:
-        msg = MIMEMultipart("alternative")
-    else:
-        msg = MIMEMultipart()
-
-    msg['From'] = addr_from
-    msg['To'] = addr_to
-    msg['Subject'] = subject
-
-    body = message
-    if not isinstance(body, bytes):
-        body = body.encode("utf-8")
-
-    body_html = """
-<html><pre style="font: monospace"><body>
-{}
-</body></pre></html>""".format(body).encode()
-
-    if html:
-        msg.attach(MIMEText(body_html, 'html', "utf-8"))
-
-    msg.attach(MIMEText(body, 'plain', "utf-8"))
-
-    # filename = "NAME OF THE FILE WITH ITS EXTENSION"
-    if filename:
-        attachment = open(filename, "rb")
-        part = MIMEBase('application', 'octet-stream')
-        part.set_payload((attachment).read())
-        encoders.encode_base64(part)
-        part.add_header('Content-Disposition', "attachment; filename= %s" % filename)
-        msg.attach(part)
-
-    # send email
-    server = smtplib.SMTP(smtp_host, smtp_port)
-    server.starttls()
-    if passwd is not None:
-        server.login(login, passwd)
-    text = msg.as_string()
-    server.sendmail(addr_from, addr_to, text)
-    server.quit()
-
-
-def __get_branch():
-    """
-    Fallback if for some reason the value vas no set by sct_launcher
-    :return:
-    """
-
-    p = subprocess.Popen(["git", "rev-parse", "--abbrev-ref", "HEAD"], stdout=subprocess.PIPE,
-                         stderr=subprocess.PIPE, cwd=__sct_dir__)
-    output, _ = p.communicate()
-    status = p.returncode
-
-    if status == 0:
-        return output.decode().strip()
-
-
-def __get_commit(path_to_git_folder=None):
-    """
-    :return: git commit ID, with trailing '*' if modified
-    """
-    if path_to_git_folder is None:
-        path_to_git_folder = __sct_dir__
-    else:
-        path_to_git_folder = abspath(path_to_git_folder)
-
-    p = subprocess.Popen(["git", "rev-parse", "HEAD"], stdout=subprocess.PIPE, stderr=subprocess.PIPE,
-                         cwd=path_to_git_folder)
-    output, _ = p.communicate()
-    status = p.returncode
-    if status == 0:
-        commit = output.decode().strip()
-    else:
-        commit = "?!?"
-
-    p = subprocess.Popen(["git", "status", "--porcelain"], stdout=subprocess.PIPE, stderr=subprocess.PIPE,
-                         cwd=path_to_git_folder)
-    output, _ = p.communicate()
-    status = p.returncode
-    if status == 0:
-        unclean = True
-        for line in output.decode().strip().splitlines():
-            line = line.rstrip()
-            if line.startswith("??"):  # ignore ignored files, they can't hurt
-                continue
-            break
-        else:
-            unclean = False
-        if unclean:
-            commit += "*"
-
-    return commit
-
-
-def __get_git_origin(path_to_git_folder=None):
-    """
-    :return: git origin url if available
-    """
-    if path_to_git_folder is None:
-        path_to_git_folder = __sct_dir__
-    else:
-        path_to_git_folder = abspath(path_to_git_folder)
-
-    p = subprocess.Popen(["git", "remote", "get-url", "origin"], stdout=subprocess.PIPE, stderr=subprocess.PIPE,
-                         cwd=path_to_git_folder)
-    output, _ = p.communicate()
-    status = p.returncode
-    if status == 0:
-        origin = output.decode().strip()
-    else:
-        origin = "?!?"
-
-    return origin
-
-
-def _git_info(commit_env='SCT_COMMIT', branch_env='SCT_BRANCH'):
-
-    sct_commit = os.getenv(commit_env, "unknown")
-    sct_branch = os.getenv(branch_env, "unknown")
-    if check_exe("git") and os.path.isdir(os.path.join(__sct_dir__, ".git")):
-        sct_commit = __get_commit() or sct_commit
-        sct_branch = __get_branch() or sct_branch
-
-    if sct_commit != 'unknown':
-        install_type = 'git'
-    else:
-        install_type = 'package'
-
-    with io.open(os.path.join(__sct_dir__, 'spinalcordtoolbox', 'version.txt'), 'r') as f:
-        version_sct = f.read().rstrip()
-
-    return install_type, sct_commit, sct_branch, version_sct
-
-
-def _version_string():
-    install_type, sct_commit, sct_branch, version_sct = _git_info()
-    if install_type == "package":
-        return version_sct
-    else:
-        return "{install_type}-{sct_branch}-{sct_commit}".format(**locals())
-
-
-def sct_progress_bar(*args, **kwargs):
-    """Thin wrapper around `tqdm.tqdm` which checks `SCT_PROGRESS_BAR` muffling the progress
-       bar if the user sets it to `no`, `off`, or `false` (case insensitive)."""
-    do_pb = os.environ.get('SCT_PROGRESS_BAR', 'yes')
-    if do_pb.lower() in ['off', 'no', 'false']:
-        kwargs['disable'] = True
-
-    return tqdm.tqdm(*args, **kwargs)
-
-
-def sct_dir_local_path(*args):
-    """Construct a directory path relative to __sct_dir__"""
-    return os.path.join(__sct_dir__, *args)
-
-
-def sct_test_path(*args):
-    """Construct a directory path relative to the sct testing data. Consults the
-    SCT_TESTING_DATA environment variable, if unset, paths are relative to the
-    current directory."""
-
-    test_path = os.environ.get('SCT_TESTING_DATA', '')
-    return os.path.join(test_path, 'sct_testing_data', *args)
-
-
-def _which_sct_binaries():
-    """
-    :return name of the sct binaries to use on this platform
-    """
-
-    if sys.platform.startswith("linux"):
-        return "binaries_linux"
-    else:
-        return "binaries_osx"
-
-
-def list2cmdline(lst):
-    return " ".join(shlex.quote(x) for x in lst)
-
-
-def run_proc(cmd, verbose=1, raise_exception=True, cwd=None, env=None, is_sct_binary=False):
-    if cwd is None:
-        cwd = os.getcwd()
-
-    if env is None:
-        env = os.environ
-
-    if sys.hexversion < 0x03000000 and isinstance(cmd, unicode):
-        cmd = str(cmd)
-
-    if is_sct_binary:
-        name = cmd[0] if isinstance(cmd, list) else cmd.split(" ", 1)[0]
-        path = None
-        binaries_location_default = sct_dir_local_path("bin")
-        for directory in (
-            sct_dir_local_path("bin"),
-        ):
-            candidate = os.path.join(directory, name)
-            if os.path.exists(candidate):
-                path = candidate
-        if path is None:
-            run_proc(["sct_download_data", "-d", _which_sct_binaries(), "-o", binaries_location_default])
-            path = os.path.join(binaries_location_default, name)
-
-        if isinstance(cmd, list):
-            cmd[0] = path
-        elif isinstance(cmd, str):
-            rem = cmd.split(" ", 1)[1:]
-            cmd = path if len(rem) == 0 else "{} {}".format(path, rem[0])
-
-    if isinstance(cmd, str):
-        cmdline = cmd
-    else:
-        cmdline = list2cmdline(cmd)
-
-    logger.debug(f"{cmdline} # in {cwd}")
-
-    shell = isinstance(cmd, str)
-
-    process = subprocess.Popen(cmd, shell=shell, cwd=cwd, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, env=env)
-    output_final = ''
-    while True:
-        # Watch out for deadlock!!!
-        output = process.stdout.readline().decode("utf-8")
-        if output == '' and process.poll() is not None:
-            break
-        if output:
-            if verbose == 2:
-                logger.debug(f"output => {output.strip()}")
-            output_final += output.strip() + '\n'
-
-    status = process.returncode
-    output = output_final.rstrip()
-
-    if status != 0 and raise_exception:
-        raise RuntimeError(output)
-
-    return status, output
-
-
-def copy_helper(src, dst, verbose=1):
-    """Copy src to dst, almost like shutil.copy
-    If src and dst are the same files, don't crash.
-    """
-    if not os.path.isfile(src):
-        folder = os.path.dirname(src)
-        contents = os.listdir(folder)
-        raise ValueError(f"Couldn't find {os.path.basename(src)} in {folder} (contents: {contents})")
-
-    try:
-        logger.info(f"cp {src} {dst}")
-        shutil.copy(src, dst)
-    except Exception as e:
-        if sys.hexversion < 0x03000000:
-            if isinstance(e, shutil.Error) and "same file" in str(e):
-                return
-        else:
-            if isinstance(e, shutil.SameFileError):
-                return
-        raise  # Must be another error
