+++ spinalcordtoolbox/utils/__init__.py
@@ -1,0 +1,4 @@
+from .fs import *
+from .fs import __version__, __sct_dir__, __data_dir__, __deepseg_dir__
+from .sys import *
+from .shell import *
